requirements.txt
# Django Core
Django==4.2.18
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-decouple==3.8

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# File Upload & Storage
Pillow==10.2.0

# Email
django-ses==3.5.2

# Task Queue
celery==5.3.6
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
